first commit
Added data structures to hold jpeg data and keep things readable.
- Markers is used to identify each byte marker in a JPEG
- Frame is what contains all data about a JPEG
- All other structs are created based on the JPEG spec.
  - https://www.w3.org/Graphics/JPEG/itu-t81.pdf

JPEG data is read only once. This requires us to know what kind of byte
we're reading as we go. I've accomplished this using stages. Each stage
says "We are currently reading a byte belonging to x", where x
is the data structure that the stage represents.

After reading in an entire Segment, we must do something with those
bytes. Since each segment is built differently, this is why the build
functions for each struct exist. Each build function takes a vector of
bytes and populates the struct accordingly.